Move echo engines from dynamo-run to dynamo-llm
That gives us some examples of how to write an engine.
