--- launch/dynamo-run/src/output/common.rs
@@ -1,32 +1,0 @@
-// SPDX-FileCopyrightText: Copyright (c) 2024-2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use std::env;
-use std::sync::LazyLock;
-use std::time::Duration;
-
-/// How long to sleep between echoed tokens.
-/// Default is 10ms which gives us 100 tok/s.
-/// Can be configured via the DYN_TOKEN_ECHO_DELAY_MS environment variable.
-pub static TOKEN_ECHO_DELAY: LazyLock<Duration> = LazyLock::new(|| {
-    const DEFAULT_DELAY_MS: u64 = 10;
-
-    let delay_ms = env::var("DYN_TOKEN_ECHO_DELAY_MS")
-        .ok()
-        .and_then(|val| val.parse::<u64>().ok())
-        .unwrap_or(DEFAULT_DELAY_MS);
-
-    Duration::from_millis(delay_ms)
-});
